refactor/#26 검증 로직 Facade 에서 수행

+++ saerojinro-api/src/main/java/goorm/saerojinro/api/review/application/ReviewFacade.java
@@ -3,21 +3,24 @@ package goorm.saerojinro.api.review.application;
 import goorm.saerojinro.api.review.presentation.request.ReviewCreateRequest;
 import goorm.saerojinro.api.review.presentation.request.ReviewUpdateRequest;
 import goorm.saerojinro.api.review.presentation.response.ReviewCreateResponse;
-import goorm.saerojinro.api.review.presentation.response.ReviewDeleteResponse;
 import goorm.saerojinro.api.review.presentation.response.ReviewListResponse;
-import goorm.saerojinro.api.review.presentation.response.ReviewUpdateResponse;
 import goorm.saerojinro.domain.lecture.application.LectureQueryService;
 import goorm.saerojinro.domain.lecture.domain.Lecture;
+import goorm.saerojinro.domain.reservation.application.ReservationQueryService;
 import goorm.saerojinro.domain.review.application.ReviewCommandService;
 import goorm.saerojinro.domain.review.application.ReviewQueryService;
 import goorm.saerojinro.domain.review.domain.Review;
+import goorm.saerojinro.domain.review.exception.ReviewNotAuthorizedException;
 import goorm.saerojinro.domain.user.application.UserQueryService;
 import goorm.saerojinro.domain.user.domain.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.time.LocalDateTime;
 import java.util.List;
+
+import static goorm.saerojinro.common.domain.BaseRole.ADMIN;
 
 @Component
 @RequiredArgsConstructor
@@ -26,6 +29,7 @@ public class ReviewFacade {
     private final ReviewCommandService reviewCommandService;
     private final LectureQueryService lectureQueryService;
     private final UserQueryService userQueryService;
+    private final ReservationQueryService reservationQueryService;
 
     @Transactional(readOnly = true)
     public ReviewListResponse getAllReview(){
@@ -46,26 +50,38 @@ public class ReviewFacade {
         User user = userQueryService.getById(request.userId());
         Lecture lecture = lectureQueryService.getByLectureId(lectureId);
 
+        if (!reservationQueryService.existsCheck(user, lecture) || LocalDateTime.now().isBefore(lecture.getEndTime())) {
+            throw new ReviewNotAuthorizedException();
+        }
+
         Review review = reviewCommandService.create(
                 user, lecture, request.content(), request.rating());
         return ReviewCreateResponse.from(review);
     }
 
     @Transactional
-    public ReviewUpdateResponse update(Long reviewId, ReviewUpdateRequest request){
+    public void update(Long reviewId, ReviewUpdateRequest request){
         User user = userQueryService.me();
 
         Review review = reviewQueryService.getByReviewId(reviewId);
-        reviewCommandService.update(user, request.content(), request.rating(), review);
-        return ReviewUpdateResponse.from(review);
+
+        if(!user.getId().equals(review.getUser().getId())){
+            throw new ReviewNotAuthorizedException();
+        }
+
+        reviewCommandService.update(request.content(), request.rating(), review);
     }
 
     @Transactional
-    public ReviewDeleteResponse delete(Long reviewId){
+    public void delete(Long reviewId){
         User user = userQueryService.me();
 
         Review review = reviewQueryService.getByReviewId(reviewId);
+
+        if(!review.getUser().getId().equals(user.getId()) && !user.getRole().equals(ADMIN)){
+          throw new ReviewNotAuthorizedException();
+        }
+
         reviewCommandService.delete(user, review);
-        return ReviewDeleteResponse.from(reviewId);
     }
 }
